Move Data.Open.Union to Data.Union
